Moved as_simd to detail
